Query numer on Dashboard
Uy 05/01/2018

+++ TechLine-ejb/src/java/entities/OrderMasterFacadeLocal.java
@@ -28,6 +28,6 @@ public interface OrderMasterFacadeLocal {
 
     String newId();
 
-    int countDoneOrder();
+    long countDoneOrder();
     
 }
